Java Loops II
Input Format:
The first line contains an integer, , denoting the number of queries.
Each line  of the  subsequent lines contains three space-separated integers describing the respective , , and  values for that query.

Constraints: 0 <=q<=500, 0<=a,b<=50, 1<= n<= 15.

Output Format:
For each query, print the corresponding series on a new line. Each series must be printed in order as a single line of  space-separated integers.